Merged PR 151: Added environment related properties
## Description

Please include a summary of the change which is being done as part of this pull request.

**Pull Request Details**:  Updated Dev server API URL in environment files.

## Type of change

Please select the relevant options. To update an options change [ ] to [x] and you would see a tick mark when the PR is actually raised.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration

- [ ] Have you unit tested your code?
- [ ] Have you ran the linters to confirm there are no linting issues?
- [ ] Have you verified the functionality aligns to the acceptance criteria?
- [ ] Have you verified if this change doesn't break any existing implementation?

## Screenshots

### desktop - Chrome Screenshot

### desktop - Edge Screenshot

Related work items: #5475