Merged PR 244: Updated web.config to serve json as static files
## Description

This change allows json files to be statically hosted. This is required for app-config.json to be returned to the application.

**Pull Request Details**:

## Type of change

Please select the relevant options. To update an options change [ ] to [x] and you would see a tick mark when the PR is actually raised.

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration

- [x] Have you unit tested your code?
- [x] Have you ran the linters to confirm there are no linting issues?
- [x] Have you verified the functionality aligns to the acceptance criteria?
- [x] Have you verified if this change doesn't break any existing implementation?

Related work items: #8199